requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
feedparser==6.0.11
FireWorks==2.0.4
Flask==3.1.2
flask-paginate==2024.4.12
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
monty==2025.3.3
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
tabulate==0.9.0
tqdm==4.67.1
tweepy==4.16.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
loguru==0.7.2
nest_asyncio==1.6.0
